+++ frontend/src/components/Agreements/AgreementEditor/AgreementEditorContext.hooks.js
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { AGREEMENT_TYPES, SERVICE_REQ_TYPES } from "../../ServicesComponents/ServicesComponents.constants";
-import { AgreementEditorContext, EditAgreementDispatchContext } from "./AgreementEditorContext";
+import { AgreementEditorContext, EditAgreementDispatchContext } from "./contexts";
 
 export const defaultState = {
     agreement: {
