+++ frontend/src/components/Agreements/AgreementEditor/contexts.js
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+export const AgreementEditorContext = createContext(null);
+export const EditAgreementDispatchContext = createContext(null);
