+++ frontend/src/components/Agreements/AgreementTypeSelect.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import Select from "../UI/Select";
+import Select from "../UI/Form/Select";
 import { useGetAgreementTypesQuery } from "../../api/opsAPI";
 import { convertCodeForDisplay } from "../../helpers/utils";
