+++ frontend/src/components/ServicesComponents/ContractTypeSelect/ContractTypeSelect.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import Select from "../../UI/Select";
+import Select from "../../UI/Form/Select";
 import { CONTRACT_TYPE_OPTIONS } from "../ServicesComponents.constants";
 
 /**
