+++ frontend/src/components/UI/Cards/BudgetCard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./BudgetCard";
