+++ frontend/src/components/UI/Cards/CurrencyCard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CurrencyCard";
