+++ frontend/src/components/UI/Cards/DocumentCard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./DocumentCard";
