+++ frontend/src/components/UI/DataViz/LineChart/readme.md
@@ -1,0 +1,3 @@
+# Data Vizualization
+
+Include React components for composing charts and graphs.
