+++ frontend/src/components/UI/DataViz/ResponsiveDonutWithInnerPercent/index.js
@@ -1,0 +1,1 @@
+export { default} from './ResponsiveDonutWithInnerPercent'
