--- frontend/src/components/UI/SummaryCard/BudgetSummaryCard/index.js
@@ -1,1 +1,0 @@
-export {default} from "./BudgetSummaryCard"
