--- frontend/src/components/UI/SummaryCard/CurrencySummaryCard/index.js
@@ -1,1 +1,0 @@
-export { default } from "./CurrencySummaryCard";
